Google Maps Integration. Untested
THIS IS A COPY OF THE STORES AND FILTERS COMMIT
I havnt built out of the latest commit because its bugged like crazy.
However I have the Google Maps started in the EditStore activity as well
as the StoresAdapter.
There have been a few other small changes throughout the project, namely
in the DatabaseManager as well as List and Stores models.

There is a non-fatal crash that happens when you load the GoogleMaps
component:
"Ensure that the "Google Maps Android API v2" is enabled." and it is
therefore untested, however the functionality is there and should work.

+++ app/src/main/java/codefactory/esy2shop/activites/EditList.java
@@ -180,19 +180,27 @@ public class EditList extends ActionBarActivity {
     }
 
     public void listSave(){
-        // Make useable DbManager
-        DatabaseManager db = new DatabaseManager(this);
-
         // Update List Object
-        list.setName(listNameField.getText().toString());
-        list.setItems(listItemsAdapter.GetListItems());
-        for(Item i : listItemsAdapter.mRemovedItems)
-        {
-            db.DeleteItem(i.getId());
-        }
-        list.setCategory(listCategoryIDs[listCategorySpinner.getSelectedItemPosition()]);
+        ArrayList<Item> saveItems = listItemsAdapter.GetListItems();
+        String saveTitle = listNameField.getText().toString();
 
-        // Apply Changes
-        list.SaveChanges(db);
+        if(saveItems.size() > 0 && !saveTitle.equals(""))
+        {
+            list.setName((saveTitle.equals("")) ? saveItems.get(0).getName() : saveTitle);
+            list.setItems(saveItems);
+            list.setCategory(listCategoryIDs[listCategorySpinner.getSelectedItemPosition()]);
+
+            // Make useable DbManager
+            DatabaseManager db = new DatabaseManager(this);
+
+            // Delete removed items
+            for(Item i : listItemsAdapter.mRemovedItems)
+            {
+                db.DeleteItem(i.getId());
+            }
+
+            // Apply Changes
+            list.SaveChanges(db);
+        }
     }
 }
